Moving the groups controller business logic to a service
